requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
django-environ==0.4.5
django-filter==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
djoser==1.7.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
